% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adasd-package.R
\docType{package}
\name{adasd-package}
\alias{adasd-package}
\title{Few R functions grouped into a package}
\description{
This package was supposted to get my exam exemption,
but it won't, due to "intuitive definition of randomness"
}

