% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x)
}
\arguments{
\item{x}{- An integer vector}
}
\value{
Most frequent value of \code{x}
}
\description{
\code{mode} returns the most frequent value of an integer vector
}
\examples{
mode(c(1,2,2))
}

