% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perms}
\alias{perms}
\title{Perms}
\usage{
perms(n)
}
\arguments{
\item{n}{- integer}
}
\value{
\code{IntegerArray(n!,n)} containing all permutations
}
\description{
\code{perms} returns all permutations of set {1,2,..,n}
}
\examples{
perms(2)
}

