% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortestpath}
\alias{shortestpath}
\title{Shortest path}
\usage{
shortestpath(G, x, y)
}
\arguments{
\item{G}{- \code{n*n} matrix with values in \{0,1\}}

\item{x}{- An integer number of starting vertex}

\item{y}{- An integer number of ending vector}
}
\value{
Length of the longest path or 999999999 if there is no path between vertices
}
\description{
\code{shortestpath} is an implementation of Dijkstra's algorithm to compute
the length of the shortest path between vertices \code{x} and \code{j} represented
by 0-1 matrix \code{G}
}
\examples{
shortestpath(matrix(c(0,1,0,0),nrow=2),1,2)
}

